task6

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ long int il;
 long long int ill;
 float f;
 
-printf("Size of a char: %d byte(s)\n", sizeof(c));
-printf("Size of an int: %d byte(s)\n", sizeof(i));
-printf("Size of a long int: %d byte(s)\n", sizeof(il));
-printf("Size of a long long int: %d byte(s)\n", sizeof(ill));
-printf("size of a float: %d byte(s)\n", sizeof(f));
+ printf("Size of a char: %d byte(s)\n", (unsigned long)sizeof(c));
+ printf("Size of an int: %d byte(s)\n", (unsigned long)sizeof(i));
+ printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(il));
+ printf("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(ill));
+ printf("size of a float: %d byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
